requirements.txt
# Core Dependencies
langchain>=0.0.350
langchain-community>=0.0.10
langchain-openai>=0.0.5

# Document Processing
PyMuPDF>=1.23.0
pdfplumber>=0.10.0
python-docx>=1.0.0

# Vector Database & Embeddings
faiss-cpu>=1.7.0
sentence-transformers>=2.2.0

# AI APIs
openai>=1.6.0
google-generativeai>=0.3.0

# Data Processing
pandas>=2.0.0
numpy>=1.21.0
tiktoken>=0.5.0

# Optional: Text-to-Speech
elevenlabs>=0.2.0
pyttsx3>=2.90

# Web Interface
Flask>=2.3.0
Werkzeug>=2.3.0
Jinja2>=3.1.0
streamlit>=1.25.0
gradio>=4.0.0

# Utilities
python-dotenv>=1.0.0
rich>=13.0.0
tqdm>=4.60.0
pydantic>=2.0.0

# Development
pytest>=7.0.0
black>=23.0.0
